+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.4] - 2021-11-12
+### Added
+- Type hint for iterable (Record)
+### Changed
+- removed None type dict entries in flexion parsing result
+### Fixed
+- minor flexion parsing improvements
+
 ## [0.8.3] - 2021-11-12
 ### Changed
 - Converted repository to [Poetry](https://python-poetry.org/) project